requirements.txt
wxpython==4.2.1
pony==0.7.19
psycopg2==2.9.10
python-dateutil==2.9.0
pyinstaller==6.12.0
attrdict2==0.0.2
pypubsub==4.0.3
transliterate==1.10.2